refactor: -optimization of cart with global states using Zustand -Better structure of components
